requirements.txt
nltk>=3.6.5
protobuf>=3.21.0
pypdf 
python-dotenv>=0.10.0
streamlit>=1.35.0
pandas>=1.5.0
faiss-cpu>=1.7.2  
PyPDF2>=1.26.0
langchain
langchain_community
langchain_core
langchain_openai
langchain_pinecone
python-jobspy
tiktoken
google-generativeai
lida
lida[transformers]
torch
openai
Tabulate 
langchain_experimental

